# Optimise line locate point