- [*] I agree to follow the project's [code of conduct](https://github.com/georust/geo/blob/main/CODE_OF_CONDUCT.md).
- [ ] I added an entry to `CHANGES.md` if knowledge of this change could be valuable to users.
---

See https://github.com/georust/geo/issues/1320

Two things here - first is optimising the LineLocatePoint algorithm. 

A new bench shows this was the original timing
```
LineString line_locate_point
                        time:   [259.34 µs 259.57 µs 259.83 µs]
                        change: [-1.6333% -1.0309% -0.5978%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
```

After updates
```
LineString line_locate_point
                        time:   [73.201 µs 73.534 µs 74.106 µs]
                        change: [-71.777% -71.684% -71.525%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
```

Great success, very nice. 

I've put some of this work out publicly under a public `ComparableDistance` trait. However I don't think that's necessarily the best idea. 

I started off just working off of Boost Geometry's `comparable_distance` idea - but I think given
 - This will be at least strictly Euclidean to start, which implies all "comparable distances" will just be the non-sqrt'ed versions which is kind of handy to know 
 - I have no idea what a `ComparableDistance` trait would look like for non-Euclidean spaces 
I think it would be better exposed as an explicit `DistanceSquared` trait that only `Euclidean` implements - if I am writing my own algo's I'd like to have it reliably convertible to a real distance. 

Keen for some feedback here - I still have some clean up here before I make this as ready to go.