@michaelkirk 2025-03-22
This seems great! No notes. 👏👏 👏 

> I think it would be better exposed as an explicit DistanceSquared trait that only Euclidean implements - if I am writing my own algo's I'd like to have it reliably convertible to a real distance.

That makes sense to me. Though I'm also not super bothered by having a ComparableDistance trait which is currently only implemented for Euclidean.